requirements.txt
fastapi
uvicorn
pydantic
python-multipart
pdfminer.six
google-generativeai
python-dotenv
